dot-dsl: Moved the solution to example.rs
