--- exercises/dot-dsl/src/graph/graph_items/mod.rs
@@ -1,2 +1,0 @@
-pub mod edge;
-pub mod node;
